Ajout de vla les tests, je pense que j'suis proche de résoudre cette connerie

+++ src/machine.h
@@ -12,15 +12,15 @@ typedef struct Station {
 } Station;
 
 typedef struct {
-    int16_t port;
-    Mac destination;
+  int16_t port;
+  Mac destination;
 } commutation;
  
 typedef struct Switch
 {
   Mac addrMac;
-  size_t nbPort;
-  size_t priorite;
+  int nbPort;
+  int priorite;
   commutation* commutations;
 } Switch;
 
@@ -35,13 +35,12 @@ typedef union Machine {
 } Machine;
 
 void initStation(Machine* machine, Mac addrMac, IP addrIp);
-void initSwitch(Machine* machine, Mac addrMac, size_t nbPorts, uint8_t priorite);
+void initSwitch(Machine* machine, Mac addrMac, int nbPorts, uint8_t priorite);
 
 void deinitMachine(Machine* machine);
 void afficheMachine(Machine* machine);
 
 
 void ajouteVoisin(Machine* machine, Mac mac);
-
 
 #endif
